feat: add support for part and supplier table and ensure determinism
This change adds support for the part and supplier tables and fixes
a bug in the distribution loading code where insertion order was not
respected.

Insertion order is important because the running weights calculation
sets the weights of each value in a given distribution based on their
order in the dists.dss file.

Our generators have now 1-1 matching outputs with Apache Trino and also
matching outputs with DuckDB except for the columns that rely on text
which is fine in this case.