feat: Introduce gzipped test fixtures and steps to reproduce. (#11)
* feat: Introduce gzipped test fixtures and steps to reproduce.

In order to build a nice test suite we need some initial fixtures
that can be easily reproduced. In this change I add two sets of data
files that I generated from the Java implementation for small scale
factors (0.01 + 0.001) the idea is to have a small set of test data
that we can easily build our test suite against.

In the change are two patch files that explain how to reproduce the data
itself (this is important since we should aim to have a reproducible
flow).