fix: Make sure table name handling is done by the crate instead of clap (#92)
* fix: Make sure table name handling is done by the crate instead of clap

To allow supporting actual short hand we have to ensure that we use
a custom TypedValueParser that is able to parse the refs correctly
and align every abbrv with its long name and corresponding enum.

I don't like this code because it's a lot of gymnastics for supporting
command flags (I feel like clap is very much too powerful for our needs
but maybe I am used to Go's stdlib provided argument handler).

There's also some non-trivial complexity in the implementation (side
effect of how clap does things).

* fix: extend TypedValueParser with possible_values

Allows us to clarify the possible values for the tables argument.

* fix: remove hidden blocks since aliases are documented